add styling for navs
