+++ package.json
@@ -85,6 +85,7 @@
     "css-loader": "^2.1.0",
     "cypress": "^3.1.2",
     "dotenv": "^6.1.0",
+    "enzyme": "^3.9.0",
     "express": "^4.16.4",
     "faker": "^4.1.0",
     "file-loader": "^3.0.1",
