+++ src/assets/styles/_components.scss
@@ -1,3 +1,5 @@
+@import 'common/Icon/styles';
+
 @import 'berths/styles';
 
 @import 'berths/InvalidSelection/styles';
