+++ src/assets/styles/_normalize.scss
@@ -1,33 +1,23 @@
-.custom-select {
-  background-color: transparent;
+body,
+span,
+textarea,
+button,
+div {
+  font-family: $font-family-base;
+  font-weight: $font-weight-normal;
 }
 
-.btn-link.btn-link {
-  color: $hel-black;
-  text-decoration: none !important;
-
-  &:active,
-  &:hover {
-    color: $hel-black !important;
-    background-color: #fff !important;
-    outline: none !important;
-    box-shadow: none !important;
-  }
+span,
+div {
+  font-size: $font-size-base;
 }
 
-.dropdown-item.active,
-.dropdown-item:active {
-  color: #000;
-  text-decoration: none;
-  background-color: #fff;
-}
-
-h1 {
-  line-height: 1;
-  margin-bottom: 2rem;
-}
-
-h3 {
-  margin-top: 2rem;
-  margin-bottom: 1rem;
+h1,
+h2,
+h3,
+h4,
+h5,
+strong {
+  font-family: $font-family-base;
+  font-weight: $font-weight-bold;
 }
