+++ src/assets/styles/main.scss
@@ -1,6 +1,6 @@
 // External style from libraries
 
-// @import './bootstrap';
+@import 'bootstrap';
 @import 'external';
 
 // Custom variables
