+++ src/components/common/Icon/_styles.scss
@@ -1,0 +1,12 @@
+.app-Icon {
+  > div {
+    height: inherit;
+    width: inherit;
+    color: inherit;
+    svg {
+      height: inherit;
+      width: inherit;
+      color: inherit;
+    }
+  }
+}
