+++ src/components/layout/Footer/_footer.scss
@@ -1,4 +1,4 @@
-.app-footer {
+.app-Footer {
   &-wrapper {
     background-color: $hel-blue;
     color: #fff;
