+++ src/components/layout/KoroSection/index.tsx
