+++ src/components/layout/Navbar/_styles.scss
@@ -1,0 +1,60 @@
+$navHeight: 3.5em;
+$navBrandWidth: 90px;
+$languageIconSize: 32px;
+$navbarTitleColor: rgba(0, 0, 0, 0.9);
+
+.app-Navbar {
+  &__top {
+    background-color: $hel-white;
+    border-bottom: 1px solid $hel-light;
+
+    .navbar {
+      padding-top: 0;
+      padding-bottom: 0;
+    }
+  }
+
+  &__bottom {
+    height: $navHeight;
+  }
+
+  &__language-dropdown {
+    .app-Icon {
+      width: $languageIconSize;
+      height: $languageIconSize;
+      color: $hel-black;
+    }
+
+    .dropdown {
+      color: $hel-black;
+      background-color: $hel-white;
+      box-shadow: none;
+
+      button {
+        display: flex;
+        align-items: center;
+        color: $hel-black;
+
+        span {
+          margin-left: 5px;
+        }
+      }
+    }
+  }
+
+  a {
+    span {
+      color: $hel-black;
+      font-weight: $display1-weight;
+    }
+
+    &:hover {
+      text-decoration: none;
+    }
+
+    .app-Icon {
+      width: $navBrandWidth;
+      color: $hel-black;
+    }
+  }
+}
