+++ src/components/layout/_styles.scss
@@ -1,3 +1,5 @@
+@import './Navbar/styles';
+
 .section-koro-top {
   &.fog:after {
     @include koro('basic', #9fc8eb, 400);
@@ -26,5 +28,42 @@
     display: block;
     margin-top: -1.25em;
     height: 1.25em;
+  }
+}
+
+.app-Layout {
+  &__hero {
+    background-image: url('../../assets/images/hero_image_berth.jpg');
+    background-size: cover;
+    background-position: right top;
+    padding: 4em 0;
+    min-height: 10em;
+    @include for-tablet {
+      min-height: 20em;
+    }
+    @include for-big-desktop {
+      min-height: 30em;
+    }
+
+    h1 {
+      font-size: 2em;
+      padding: 8px 16px;
+
+      @include for-phone {
+        font-size: 3em;
+      }
+      @include for-desktop {
+        font-size: 4em;
+      }
+      color: $hel-white;
+      text-shadow: 1px 1px 4px rgba(0, 0, 0, 0.2);
+    }
+  }
+
+  &__koro {
+    > div {
+      margin-top: 3em;
+      margin-bottom: 4em;
+    }
   }
 }
