+++ src/components/layout/index.tsx
@@ -1,0 +1,61 @@
+import React, { Fragment } from 'react';
+import { FormattedHTMLMessage, FormattedMessage, InjectedIntlProps, injectIntl } from 'react-intl';
+import { Col, Container, Row } from 'reactstrap';
+import Footer from './Footer';
+import KoroSection from './KoroSection';
+import Navbar from './Navbar';
+
+type Props = {
+  children: any;
+  hero?: boolean;
+} & InjectedIntlProps;
+
+const getHeroContentLink = (locale: string) => {
+  switch (locale) {
+    case 'en':
+      return 'https://www.hel.fi/helsinki/en/culture/recreation/boating/boat-berths/';
+    case 'sv':
+      return 'https://www.hel.fi/helsinki/sv/kultur-och-fritid/friluftsliv/botliv/batplatser/';
+    default:
+      return 'https://www.hel.fi/helsinki/fi/kulttuuri-ja-vapaa-aika/ulkoilu/veneily/kaupungin-venepaikat/venepaikan-hakeminen/';
+  }
+};
+
+const Layout = ({ children, hero, intl: { locale } }: Props) => (
+  <div className="app-Layout">
+    <Navbar />
+
+    {hero && (
+      <div className="app-Layout__hero">
+        <Container>
+          <FormattedMessage tagName="h1" id="site.title" />
+        </Container>
+      </div>
+    )}
+
+    <KoroSection bottom color="white" className="app-Layout__koro">
+      <Container>
+        <Row>
+          <Col lg={{ size: 10, offset: 1 }} xl={{ size: 8, offset: 2 }}>
+            <FormattedMessage tagName="h1" id="hero.title" />
+            <FormattedMessage tagName="p" id="hero.paragraph.first" />
+            <FormattedHTMLMessage
+              tagName="p"
+              id="hero.paragraph.second"
+              values={{ url: getHeroContentLink(locale) }}
+            />
+          </Col>
+        </Row>
+      </Container>
+    </KoroSection>
+
+    <KoroSection color="fog" bottom>
+      {children}
+    </KoroSection>
+    <KoroSection bottom color="blue">
+      <Footer />
+    </KoroSection>
+  </div>
+);
+
+export default injectIntl(Layout);
