+++ src/components/pages/SelectedBerthPage.tsx
@@ -6,7 +6,7 @@ import { getBerthFilterByValues } from '../../utils/berths';
 import SelectedBerths from '../berths/SelectedBerths';
 import Icon from '../common/Icon';
 import LocalizedLink from '../common/LocalizedLink';
-import Layout from '../layout/Layout';
+import Layout from '../layout';
 import SelectedBerthsLegend from '../legends/SelectedBerthsLegend';
 
 import { BoatTypes } from '../../types/boatTypes';
